pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ukcat"
dynamic = ["version"]
description = 'This repository is the home of the UK Charity Activity Tags, a project to classify every UK registered charity using two classification taxonomies.'
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
keywords = []
authors = [{ name = "David Kane", email = "david@dkane.net" }]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
  "airtable-python-wrapper",
  "pandas",
  "numpy",
  "scikit-learn",
  "nltk",
  "python-dotenv",
  "python-slugify",
  "click",
  "tqdm",
  "jinja2",
]

[project.optional-dependencies]
docs = [
  "mkdocs",
  "mkdocs-material",
  "mkdocs-blog-plugin",
  "mkdocs-git-revision-date-localized-plugin",
  "mkdocs-rss-plugin",
]
notebooks = ["ipykernel", "jupyter", "jupyterlab"]
lint = ["mypy>=1.0.0", "ruff"]

[project.urls]
Homepage = "https://www.charityclassification.org.uk/"
Documentation = "https://github.com/charity-classification/ukcat#readme"
Issues = "https://github.com/charity-classification/ukcat/issues"
Source = "https://github.com/charity-classification/ukcat"

[project.scripts]
ukcat = "ukcat.cli:ukcat"

[tool.hatch.version]
path = "src/ukcat/__about__.py"

[tool.hatch.envs.default]
dependencies = ["coverage[toml]>=6.5", "pytest"]
[tool.hatch.envs.default.scripts]
test = "pytest {args:tests}"
test-cov = "coverage run -m pytest {args:tests}"
cov-report = ["- coverage combine", "coverage report"]
cov = ["test-cov", "cov-report"]

[tool.hatch.envs.docs]
features = ["docs"]

[tool.hatch.envs.docs.scripts]
serve = "mkdocs serve"
build = "mkdocs build"
deploy = "mkdocs gh-deploy --force"

[tool.hatch.envs.lint]
detached = true
features = ["lint"]

[tool.hatch.envs.lint.scripts]
typing = "mypy --install-types --non-interactive {args:src/ixbrlparse tests}"
style = ["ruff check {args:.}", "ruff format --check --diff {args:.}"]
fmt = ["ruff format {args:.}", "ruff check --fix {args:.}", "style"]
all = ["style", "typing"]

[tool.mypy]
ignore_missing_imports = true
mypy_path = "$MYPY_CONFIG_FILE_DIR/stubs/"
files = "./stubs/**.*"

[tool.ruff]
target-version = "py38"
line-length = 120

[tool.coverage.run]
source_pkgs = ["ukcat", "tests"]
branch = true
parallel = true
omit = ["src/ukcat/__about__.py"]

[tool.coverage.paths]
ukcat = ["src/ukcat", "*/ukcat/src/ukcat"]
tests = ["tests", "*/ukcat/tests"]

[tool.coverage.report]
exclude_lines = ["no cov", "if __name__ == .__main__.:", "if TYPE_CHECKING:"]
